feat: move storage to UserDefaults
