+++ p2p_wallet/Resources/AppDelegate/AppDelegate+Injection.swift
@@ -72,11 +72,14 @@ extension Resolver: ResolverRegistering {
             .implements(PincodeStorageType.self)
             .implements(AccountStorageType.self)
             .implements(PincodeSeedPhrasesStorage.self)
-            .implements(SendViaLinkStorageType.self)
             .implements((AccountStorageType & NameStorageType).self)
             .implements((AccountStorageType & PincodeStorageType & NameStorageType).self)
             .implements((ICloudStorageType & AccountStorageType & NameStorageType).self)
             .implements((ICloudStorageType & AccountStorageType & NameStorageType & PincodeStorageType).self)
+            .scope(.application)
+        
+        register { SendViaLinkStorageImpl() }
+            .implements(SendViaLinkStorage.self)
             .scope(.application)
 
         // API Gateway
