+++ p2p_wallet/Services/Storage/KeychainStorage.swift
@@ -22,8 +22,6 @@ class KeychainStorage: StorageType {
 
     let iCloudAccountsKey = "Keychain.Accounts"
     private let iCloudKeychainToLocalKeychainMigrated = "iCloudKeychainToLocalKeychainMigrated"
-    
-    let iCloudSendViaLinkSeedsKey = "Keychain.iCloudSendViaLinkSeedsKey"
 
     // MARK: - Properties
